pyproject.toml
[tool.poetry]
name = "centralize-backend"
version = "0.1.0"
description = ""
authors = ["Cândido Júnior <candidonogueirajr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-split-settings = "^1.3.1"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
django-mptt = "^0.16.0"
pillow = "^10.3.0"
model-bakery = "^1.18.2"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
